rename various core files in line with refactored classes they contain.
